refactor(suite): move handler related functions to correct files
